// packages/twenty-apps/wyre/linkedin-voyager-shim/src/domain/unipile/unipile-message-received-webhook.type.ts
import { type UnipileWebhookAttendee } from 'src/domain/unipile/unipile-webhook-attendee.type';

export type UnipileMessageReceivedWebhookPayload = {
  event: 'message_received';
  account_id: string;
  account_type: string;
  chat_id: string;
  message_id: string;
  message: string | null;
  timestamp: string;
  sender: UnipileWebhookAttendee;
  attendees: UnipileWebhookAttendee[];
  account_info: { user_id: string };
};